refactor: use modular models package
