+++ app.py
@@ -5,8 +5,7 @@ Thin runner that uses the unified factory and runs Socket.IO.
 import re, datetime as dt
 from flask import Blueprint, request, jsonify
 from flask_login import LoginManager, login_user, logout_user, login_required, current_user
-from app.db import db as db
-from app.models import User
+from app.models import db, User
 from flask_socketio import SocketIO
 from app import create_app
