+++ app/models/__init__.py
@@ -1,14 +1,15 @@
 from .base import db, Model, metadata
 
 # Import model modules so tables register with metadata
-from .characters import Character          # noqa: F401
-from .items import Item                    # noqa: F401
+from .users import User                      # noqa: F401
+from .characters import Character            # noqa: F401
+from .items import Item                      # noqa: F401
 from .inventory import ItemInstance, CharacterInventory   # noqa: F401
-# from .crafting import Recipe             # noqa: F401
+# from .crafting import Recipe               # noqa: F401
 
 __all__ = [
     "db", "Model", "metadata",
-    "Character",
+    "User", "Character",
     "Item", "ItemInstance", "CharacterInventory",
     # "Recipe",
 ]
