+++ app/models/users.py
@@ -1,14 +1,14 @@
 import uuid, datetime as dt
 from flask_login import UserMixin
 
-from ..db import db as db
+from .base import db, Model
 
 
 def gen_uuid() -> str:
     return str(uuid.uuid4())
 
 
-class User(db.Model, UserMixin):
+class User(Model, UserMixin):
     __tablename__ = "users"
 
     user_id = db.Column(db.String, primary_key=True, default=gen_uuid)
